chore(shared): move commander to core package
