+++ packages/core/compiled/commander/license
