+++ packages/core/compiled/commander/package.json
