+++ packages/core/compiled/commander/typings/index.d.ts
